+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/BulkErrorCrawlerIT.java
@@ -33,7 +33,7 @@ public class BulkErrorCrawlerIT extends AbstractIT {
     slugs.add("boxes");
     slugs.add("instruments");
     slugs.add("kitdescriptors");
-    slugs.add("indices");
+    slugs.add("indexfamily/list");
     slugs.add("studies");
     slugs.add("printers");
     slugs.add("experiments");