+++ miso-web/src/main/webapp/scripts/list_indices.js
@@ -24,29 +24,61 @@
 ListTarget.index = {
   name: "Indices",
   createUrl: function(config, projectId) {
-    return "/miso/rest/indices/dt" + (config.platformType ? "/platform/" + config.platformType : "");
+    return config.platformType ? Urls.rest.indices.platformDatatable(config.platformType) : Urls.rest.indices.datatable;
   },
   getQueryUrl: null,
   createBulkActions: function(config, projectId) {
     var actions = [];
+    if (config.indexFamilyId && config.isAdmin) {
+      actions.push({
+        name: 'Edit',
+        action: function(items) {
+          window.location = Urls.ui.indices.bulkEdit + '?' + jQuery.param({
+            ids: items.map(Utils.array.getId).join(',')
+          });
+        }
+      }, ListUtils.createBulkDeleteAction('Indices', 'indices', function(index) {
+        return index.name + ' (' + index.sequence + ')';
+      }));
+    }
     if (config.additionalBulkActions) {
       actions = actions.concat(config.additionalBulkActions);
     }
     return actions;
   },
   createStaticActions: function(config, projectId) {
-    return [];
+    return !config.indexFamilyId || !config.isAdmin ? [] : [{
+      name: 'Add',
+      handler: function() {
+        Utils.showDialog('Create Indices', 'Create', [{
+          property: 'quantity',
+          type: 'int',
+          label: 'Quantity',
+          required: true,
+          value: 1
+        }], function(result) {
+          if (result.quantity < 1) {
+            Utils.showOkDialog('Create Indices', ["Quantity must be 1 or more."]);
+            return;
+          }
+          window.location = Urls.ui.indices.bulkCreate + '?' + jQuery.param({
+            indexFamilyId: config.indexFamilyId,
+            quantity: result.quantity,
+          });
+        });
+      }
+    }];
   },
   createColumns: function(config, projectId) {
     return [{
       "sTitle": "Platform",
-      "include": !config.platformType,
+      "include": !config.platformType && !config.indexFamilyId,
       "iSortPriority": 3,
       "mData": "family.platformType",
       "mRender": ListUtils.render.platformType
     }, {
       "sTitle": "Family",
-      "include": true,
+      "include": !config.indexFamilyId,
       "iSortPriority": 2,
       "mData": "family.name"
     }, {
@@ -55,10 +87,18 @@ ListTarget.index = {
       "iSortPriority": 1,
       "mData": "name"
     }, {
-      "sTitle": "Sequence",
+      "sTitle": "Sequence(s)",
       "include": true,
       "iSortPriority": 0,
-      "mData": "sequence"
+      "bSortable": false,
+      "mData": "sequence",
+      "mRender": function(data, type, full) {
+        if (type === 'display' && full.family.fakeSequence) {
+          return full.realSequences.join(', ');
+        } else {
+          return data;
+        }
+      }
     }, {
       "sTitle": "Position",
       "include": true,